+++ tests/fscript_test.cc
@@ -1171,9 +1171,9 @@ TEST(FExpr, not_eq) {
 TEST(FScript, global_func) {
   value_t v;
   tk_object_t* obj = object_default_create();
-  fscript_register_func("foo", func_foo);
+  fscript_register_func("foobar", func_foo);
 
-  fscript_eval(obj, "foo()", &v);
+  fscript_eval(obj, "foobar()", &v);
   ASSERT_EQ(value_int(&v), 123);
   value_reset(&v);
 
@@ -2632,6 +2632,46 @@ TEST(FScript, func_no_args) {
                "{return a + b;}; var c = hello(foo1(), bar());c",
                &v);
   ASSERT_EQ(value_int(&v), 3);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
+TEST(FScript, module_str) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj,
+"\
+var a = require_str('\
+function foo(a, b) {\
+  return a + b;\
+}\
+')\
+var b = a.foo(100, 99);\
+b\
+",
+               &v);
+  ASSERT_EQ(value_int(&v), 199);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
+TEST(FScript, module_file) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj,
+"\
+var a = require('./tests/fscripts/foo.fs')\
+var b = a.foo(100, 99);\
+var c = a.bar(100, 99);\
+var d = b + c;\
+d\
+",
+               &v);
+  ASSERT_EQ(value_int(&v), 200);
   value_reset(&v);
 
   TK_OBJECT_UNREF(obj);